# Conflicts:
#	Assets/Resources/Rules/dragonDefinitions.xml
#	Docs/Content/HungryDragonContent_Dragons.xlsx


Former-commit-id: 84170ceb29291920dc84c2faa5ad4bf71259f443

+++ Assets/Game/Model/MissionsAndEvents/Objectives/TrackingObjectiveBase.cs
@@ -44,7 +44,7 @@ public abstract class TrackingObjectiveBase {
 	}
 
 	public float progress { 
-		get { return m_targetValue > 0f ? m_tracker.currentValue/targetValue : 1f; }	// Protect for infinite values!
+		get { return m_targetValue > 0f ? (float)((double)m_tracker.currentValue/(double)targetValue) : 1f; }	// Protect for infinite values!
 	}
 
 	// State
